requirements.txt
alembic==1.7.6
autopep8==1.6.0
click==8.0.3
colorama==0.4.4
Flask==2.0.2
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
install==1.3.5
itsdangerous==2.0.1
Jinja2==3.0.3
Mako==1.1.6
MarkupSafe==2.0.1
psycopg2==2.9.3
pycodestyle==2.8.0
python-dotenv==0.19.2
SQLAlchemy==1.4.31
toml==0.10.2
Werkzeug==2.0.3
